fix: add clippy allow for cross-platform statvfs casts
macOS has u32 f_blocks/f_bavail fields while Linux has u64.
Casts are necessary for macOS but clippy warns on Linux.

+++ aura-daemon/src/collectors/disk.rs
@@ -133,9 +133,13 @@ fn get_fs_stats(mountpoint: &[u8]) -> (u64, u64, u64, f32) {
     }
 
     let s = unsafe { s.assume_init() };
-    let frsize = s.f_frsize;
-    let total = s.f_blocks.saturating_mul(frsize);
-    let available = s.f_bavail.saturating_mul(frsize);
+    // Cast to u64 for cross-platform compatibility: macOS has u32 fields
+    #[allow(clippy::unnecessary_cast)]
+    let frsize = s.f_frsize as u64;
+    #[allow(clippy::unnecessary_cast)]
+    let total = (s.f_blocks as u64).saturating_mul(frsize);
+    #[allow(clippy::unnecessary_cast)]
+    let available = (s.f_bavail as u64).saturating_mul(frsize);
     let used = total.saturating_sub(available);
     let percent = if total > 0 {
         (used as f32 / total as f32) * 100.0
